Create 037. Armstrong check
Q37. WAPC to input a positive integer. Assume that the number is of 3 digits. Check and print if the number is Armstrong or not. Note: An Armstrong number is a number that is equal to the sum of its own digits each raised to the power of the number of digits, for example, 153 = 1³ + 5³ + 3³.